Move GPU class to a more proper package.
